Health Bar, Intro Banner, More Dancers
I added in a simple health bar that will display how close you are do death by not dancing enough.

There is a really crappy banner that flashes on screen before you start.

After 30 seconds of playing the back row will jump in. This would be nicer if we spaced it out by adding 2 at a time.